support MG pod deployment

+++ docker/edgemicro/entrypoint.sh
@@ -7,22 +7,9 @@ exec 2>&1
 
 echo "Log Location should be: [ $LOG_LOCATION ]"
 
-#echo $EDGEMICRO_DECORATOR >> /tmp/test.txt
-#echo $EDGEMICRO_ORG >> /tmp/test.txt
-#echo $EDGEMICRO_ENV >> /tmp/test.txt
-#echo $EDGEMICRO_KEY >> /tmp/test.txt
-#echo $EDGEMICRO_SECRET >> /tmp/test.txt
-#echo $EDGEMICRO_MGMTURL >> /tmp/test.txt
-#echo $EDGEMICRO_ADMINEMAIL >> /tmp/test.txt
-#echo $EDGEMICRO_ADMINPASSWORD >> /tmp/test.txt
-#echo $POD_NAME >> /tmp/test.txt
-#echo $POD_NAMESPACE >> /tmp/test.txt
-#echo $INSTANCE_IP >> /tmp/test.txt
 SERVICE_NAME=`echo "${SERVICE_NAME}" | tr '[a-z]' '[A-Z]'`
-#echo $SERVICE_NAME >> /tmp/test.txt
 SERVICE_PORT_NAME=${SERVICE_NAME}_SERVICE_PORT
 SERVICE_PORT=${!SERVICE_PORT_NAME}
-#echo $SERVICE_PORT >> /tmp/test.txt
 proxy_name=edgemicro_$POD_NAME
 product_name=$proxy_name-product
 
@@ -32,7 +19,9 @@ if [ ${EDGEMICRO_CONFIG} != "" ]; then
 	echo ${EDGEMICRO_CONFIG} | base64 --decode > /opt/apigee/.edgemicro/$EDGEMICRO_ORG-$EDGEMICRO_ENV-config.yaml
 	# Decorate Proxy with the proxy name
   sed -i.bak s/proxy_name/${proxy_name}/g /tmp/proxies.yaml
-  sed -i.bak '/edgemicro:/r /tmp/proxies.yaml' /opt/apigee/.edgemicro/$EDGEMICRO_ORG-$EDGEMICRO_ENV-config.yaml
+  if [ ${EDGEMICRO_DECORATOR} != "" ]; then
+         sed -i.bak '/edgemicro:/r /tmp/proxies.yaml' /opt/apigee/.edgemicro/$EDGEMICRO_ORG-$EDGEMICRO_ENV-config.yaml
+  fi
   chown apigee:apigee /opt/apigee/.edgemicro/*
 fi
 
@@ -52,11 +41,12 @@ su - apigee -m -c "$commandString"
 my_handler() {
   echo "my_handler" >> /tmp/entrypoint.log
   su - apigee -m -c "cd /opt/apigee && edgemicro stop"
-  #Attempt deleting the proxy here
-  curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apiproducts/${product_name}
-  curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/environments/${EDGEMICRO_ENV}/apis/${proxy_name}/revisions/1/deployments
-  curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apis/${proxy_name}
-
+  if [ ${EDGEMICRO_DECORATOR} != "" ]; then
+      #Attempt deleting the proxy here
+      curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apiproducts/${product_name}
+      curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/environments/${EDGEMICRO_ENV}/apis/${proxy_name}/revisions/1/deployments
+      curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apis/${proxy_name}
+  fi
   #edgemicro stop
 }
 
@@ -64,11 +54,11 @@ my_handler() {
 term_handler() {
   echo "term_handler" >> /tmp/entrypoint.log
   su - apigee -m -c "cd /opt/apigee && edgemicro stop"
-
-  curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apiproducts/${product_name}
-  curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/environments/${EDGEMICRO_ENV}/apis/${proxy_name}/revisions/1/deployments
-  curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apis/${proxy_name}
-
+  if [ ${EDGEMICRO_DECORATOR} != "" ]; then
+      curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apiproducts/${product_name}
+      curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/environments/${EDGEMICRO_ENV}/apis/${proxy_name}/revisions/1/deployments
+      curl -v -X DELETE -u $EDGEMICRO_ADMINEMAIL:$EDGEMICRO_ADMINPASSWORD -H "Content-Type:application/x-www-form-urlencoded" ${EDGEMICRO_MGMTURL}/v1/organizations/${EDGEMICRO_ORG}/apis/${proxy_name}
+  fi
   #edgemicro stop
   exit 143; # 128 + 15 -- SIGTERM
 }
